+++ jaorm-processor/src/main/java/io/github/ulisse1996/jaorm/processor/validation/impl/QueryValidator.java
@@ -14,9 +14,18 @@ import io.github.ulisse1996.jaorm.specialization.SingleKeyDao;
 import io.github.ulisse1996.jaorm.specialization.TripleKeyDao;
 
 import javax.annotation.processing.ProcessingEnvironment;
-import javax.lang.model.element.*;
+import javax.lang.model.element.Element;
+import javax.lang.model.element.ExecutableElement;
+import javax.lang.model.element.Name;
+import javax.lang.model.element.TypeElement;
+import javax.lang.model.element.VariableElement;
 import javax.lang.model.type.TypeKind;
-import java.util.*;
+import java.util.Arrays;
+import java.util.Collections;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Optional;
+import java.util.Set;
 import java.util.stream.Collectors;
 
 public class QueryValidator extends Validator {
@@ -125,7 +134,7 @@ public class QueryValidator extends Validator {
     }
 
     private void checkSpecs(String sql, ExecutableElement method) {
-        if (sql.toUpperCase().startsWith("SELECT")) {
+        if (isValidSelect(sql)) {
             checkReturnMethod(method);
             return;
         } else if (sql.toUpperCase().startsWith("DELETE") || sql.toUpperCase().startsWith("UPDATE")) {
@@ -134,6 +143,10 @@ public class QueryValidator extends Validator {
         }
 
         throw new ProcessorException(String.format("Operation not supported for sql %s in method %s", sql, method));
+    }
+
+    private boolean isValidSelect(String sql) {
+        return sql.toUpperCase().startsWith("SELECT") || sql.toUpperCase().startsWith("WITH");
     }
 
     private void checkReturnMethod(ExecutableElement method) {