+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 - Feature: Distinct for DSL SubQuery
 - Feature: Simple EnumConverter
 - Feature: InlineValue with Alias
+- Feature: Support WITH for Query Annotation
 - Removal: Removed notLike and like
 - Refactor: Refactor of IT Tests
 - Bugfix: Fix missing upsert for entities with relationships